chore: add yopa container to project
yopa is an acronym for Your Own Personal AWS. It is a project that aims
to simulate AWS services on a container for testing purposes.

This commit adds the `yopa` entry on the project's `docker-compose` file
and the `yopa-config.yml` file to hold YOPA settings.

This commit also updates the `docker-compose` version to version `2.1`
so we can use docker-compose's healthcheck system to check a container's
health.