chore: use postgresql instead of sqlite
Use PostgreSQL as the project's database engine, instead of SQLite. This
change was expected since the project's genesis.

This commit adds to the project a migration script, a config file for
the database and a project setup script.

The sqlite dependencies have been removed from the project and
postgresql dependencies for sequelize have been added.