config: refactor getter
The `getConfig` function returned a function that was used to get
environment specific configuration. Since that behaviour was never used,
this commit changes the `getConfig` function to, instead, return the
NODE_ENV environment configuration, with fallback to the 'development'
environment.

Custom environment reading is still accepted, but as a second argument
of the `getConfig` function.